Updated the loss tangent to be more representative
This updates the loss tangent values for the materials based on data from #8

I haven't changed any of the dielectric constants, but I have updated
the stackup definitions to properly use the material features. There was a
bug that was causing the stackup definitions to use the defaults in JSL (which are also bad).